requirements.txt
begins
click
pandas
pybedtools
pysam
pyvcf